Improved packaging
This commits has 2 main goals:

- Allow to provide vconf, vconf-devel, vconf-keys-devel
for solving issue when building.

- that done, error came because the vconftool replacement
coulded not run without a server buxton what is a problem
at build time. Thus the tool is replaced using an agnostic
tool written in bash.

There is an issue with that current version: it provides a
package that it also obsoletes. This issue will be fixed soon.

Change-Id: I9ea93d61cfe12fa23448fd6367d372cc7bbf55d9
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>